[TASK-32] Restructure folders
