+++ src/helpers/api.js
@@ -1,6 +1,6 @@
 import axios from 'axios';
 
-import UnauthorizedInterceptor from 'utils/unauthorizedInterceptor';
+import UnauthorizedInterceptor from 'helpers/unauthorizedInterceptor';
 
 const instance = axios.create({
   baseURL: process.env.REACT_APP_BACKEND_URL,
